+++ core/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Added
+
+- Implement `FromStr` for `ObjectUuid` and `ServiceUuid`.
+
+
 # 0.1.0 (November 24th, 2023)
 
 - Initial release.